+++ web/src/settings/SettingsArea.tsx
@@ -139,7 +139,7 @@ export class SettingsArea extends React.Component<Props, State> {
                             )}
                         />
                         <Route
-                            path={`${this.props.match.url}/accounts`}
+                            path={`${this.props.match.url}/external-accounts`}
                             key="hardcoded-key" // see https://github.com/ReactTraining/react-router/issues/4578#issuecomment-334489490
                             exact={true}
                             // tslint:disable-next-line:jsx-no-lambda