+++ CHANGELOG.md
@@ -60,6 +60,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 * Added `nameIDFormat` in the `saml` auth provider to set the SAML NameID format. The default changed from transient to persistent.
 * Experimental env var expansion in site config JSON: set `SOURCEGRAPH_EXPAND_CONFIG_VARS=1` to replace `${var}` or `$var` (based on environment variables) in any string value in site config JSON (except for JSON object property names).
 * The new (optional) SAML `serviceProviderIssuer` site config property (in an `auth.providers` array entry with `{"type":"saml", ...}`) allows customizing the SAML Service Provider issuer name.
+* The site admin area now has an "Auth" section that shows the enabled authentication provider(s) and users' external accounts.
 
 ## 2.7.6